+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.3] - 2024-04-10
+
+### Fixed
+- Suppressed DEBUG logging messages in production builds
+- Changed directory not found warnings to debug level messages for plugins and templates directories
+- Enforced minimum log level of INFO in production builds
+
 ## [1.9.2] - 2024-04-10
 
 ### Fixed